==> Script/03_SpatialData_01_CCA.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#1. 
##1.1. 
```{r}
library(Seurat)
library(ggplot2)
library(data.table)
library(magrittr)
library(plyr)
library(dplyr)
library(ggpubr)
library(ggrepel)
library(ggridges)
```
##1.2.
```{r}
#
Uninfect <- readRDS("../01_Data/03_01_Uninfect.Rds")
Uninfect@assays$Spatial@data <- Uninfect@assays$Spatial@counts
#
Infect <- readRDS("../01_Data/03_01_Infect.Rds")
Infect@assays$Spatial@data <- Infect@assays$Spatial@counts
#
Infect_IfngrKO <- readRDS("../01_Data/03_01_Infect_Ifngr1KO.Rds")
Infect_IfngrKO@assays$Spatial@data <- Infect_IfngrKO@assays$Spatial@counts
```
##1.3. 
```{r}
Uninfect$orig.ident <- "WT"
Infect$orig.ident <- "Infect"
Infect_IfngrKO$orig.ident <- "Infect_IfngrKO"
#
sce.big <- list(Uninfect,Infect,Infect_IfngrKO)
names(sce.big) <- c("WT","Infect","Infect_IfngrKO")
```

#2. 
##2.1. 
```{r}
PlotData <- do.call(rbind,list(sce.big$WT@meta.data,sce.big$Infect@meta.data,sce.big$Infect_IfngrKO@meta.data))
PlotData$orig.ident <- factor(PlotData$orig.ident,levels = c("WT","Infect","Infect_IfngrKO"))
#
p1 <- ggplot()+
  geom_density(data = PlotData,aes(x=nCount_Spatial,group=orig.ident,color=orig.ident),size=1)+
  geom_density(data = PlotData,aes(x=nCount_Spatial),color="red",size=1,linetype=2)+
  theme_bw()+
  theme(panel.grid = element_blank())+
  geom_vline(xintercept = c(1000,5000))+
  scale_color_manual(values = c("#E4B54D", "#82B368", "#AEE6D6"))
#
p2 <- ggplot()+
  geom_density(data = PlotData,aes(x=nFeature_Spatial,group=orig.ident,color=orig.ident),size=1)+
  geom_density(data = PlotData,aes(x=nFeature_Spatial),color="red",size=1,linetype=2)+
  theme_bw()+
  theme(panel.grid = element_blank())+
  geom_vline(xintercept = c(600,2500))+
  scale_color_manual(values = c("#E4B54D", "#82B368", "#AEE6D6"))

```
##2.2. 
```{r,fig.width=12,fig.height=4}
p1+p2
```

#3. 
```{r}
sce.big <- lapply(sce.big,function(x){
  x <- subset(x,nCount_Spatial>1000&nCount_Spatial<5000)
  x <- subset(x,nFeature_Spatial>600&nFeature_Spatial<2500)
  return(x)
})
#
sce.big <- lapply(sce.big,function(x){
x <- SCTransform(x, assay = "Spatial", return.only.var.genes = FALSE, verbose = FALSE)
return(x)
    }
)

features <- SelectIntegrationFeatures(object.list = sce.big, nfeatures = 2000)
sce.big  <- PrepSCTIntegration(object.list = sce.big, anchor.features = features)
sce.big  <- FindIntegrationAnchors(object.list = sce.big, normalization.method = "SCT", anchor.features = features)
sce.big  <- IntegrateData(anchorset = sce.big, normalization.method = "SCT")

DefaultAssay(sce.big) <- "integrated"
sce.big <- RunPCA(sce.big, npcs = 20, verbose = FALSE)
sce.big <- RunUMAP(sce.big, reduction = "pca", dims = 1:10)
sce.big <- FindNeighbors(sce.big, reduction = "pca", dims = 1:10)
sce.big <- FindClusters(sce.big, resolution = seq(.1,1,.1),verbose = TRUE)
sce.big$orig.ident <- factor(sce.big$orig.ident,levels = c("WT","Infect","Infect_IfngrKO"))
names(sce.big@images) <- c("WT","Infect","Infect_IfngrKO")
DefaultAssay(sce.big) <- "SCT"

#
saveRDS(sce.big,"../01_Data/03_01_CCA_Allsample.Rds")
```

#4.
```{r,fig.width=7,fig.height=4}
PlotData <- data.frame(
  Group = rep(table(sce.big$orig.ident)%>%names(),2),
  Number = c(table(sce.big$orig.ident)%>%as.numeric(),dim(Uninfect)[2],dim(Infect)[2],dim(Infect_IfngrKO)[2]),
  Type = rep(c("After QC","Before QC"),c(3,3))
)
PlotData$Group <- factor(PlotData$Group,levels = c("WT","Infect","Infect_IfngrKO"))
PlotData$Type <- factor(PlotData$Type,levels = c("Before QC","After QC"))
setorder(PlotData,Group)
setorder(PlotData,Type)
#
ggplot(PlotData)+
  geom_bar(aes(x=Group,y=Number,fill=Type),stat = "identity",position = "dodge")+
  geom_text(aes(x=Group,y=Number,label=Number),position = position_dodge2(.9),vjust=-.1,size=5)+
  scale_fill_manual(values = c("#619CFF","#F8766D"))+theme_bw()+
  xlab("")+ylab("Spot numbers")+theme(panel.grid = element_blank())
```
