==> Script/03_SpatialData_02_SubRegion.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#1. 
##1.1. 
```{r}
library(Seurat)
library(ggplot2)
library(data.table)
library(magrittr)
library(ggpubr)
library(ggrepel)
library(SeuratDisk)
library(ggridges)
library(enrichplot)
library(clusterProfiler)
library(dplyr)
library(plyr)
library(org.Mm.eg.db)
library(ggalluvial)
library(Startrac)
library(ComplexHeatmap)
library(RColorBrewer)
```
##1.2. 
```{r,fig.width=12,fig.height=4}
sce.big <- readRDS("../01_Data/03_01_CCA_Allsample.Rds")
DefaultAssay(sce.big) <- "SCT"
```
##1.3.
```{r}
sce.big@meta.data%<>%mutate(
  Region = case_when(
    integrated_snn_res.0.2 == 0 ~"Region1",
    integrated_snn_res.0.2 == 1 ~"Region2",
    integrated_snn_res.0.2 == 2 ~"Region3",
    integrated_snn_res.0.2 == 3 ~"Region4",
    integrated_snn_res.0.2 == 4 ~"Region5",
    integrated_snn_res.0.2 == 5 ~"Region6",
    integrated_snn_res.0.2 == 6 ~"Region7",
    integrated_snn_res.0.2 == 7 ~"Region8",
    integrated_snn_res.0.2 == 8 ~"Region9",
    integrated_snn_res.0.2 == 9 ~"Region10"
  )
)
sce.big$Region <- factor(sce.big$Region,levels = c("Region1","Region2","Region3","Region4","Region5","Region6","Region7","Region8","Region9","Region10"))
Idents(sce.big) <- "Region"
```

#2. Plot
##2.1. 
```{r,fig.width=25,fig.height=10}
SpatialDimPlot(sce.big,group.by = "Region",label = F,pt.size.factor = 5,ncol = 3,image.alpha = 0)&
  scale_fill_manual(values =  c("#C497B2","#D8383A","#14517C", "#E7EFFA", "#96C37D", "#F3D266", "#2F7FC1", "#F8F3F9","#A9B8C6","#F7E1ED"))&
  theme(plot.title =  element_text(size = 25),legend.title = element_text(size = 15),legend.text = element_text(size = 15))&
  guides(fill=guide_legend(override.aes = list(size=8)))
```
##2.2. 
###2.2.1. 
```{r,fig.width=4,fig.height=5}
plot_data <- table(sce.big$orig.ident,sce.big$Region)%>%as.data.frame()
colnames(plot_data) <- c("group","X","value")
plot_data$group <- factor(plot_data$group,levels = c("WT","Infect","Infect_IfngrKO"))
plot_data%<>%as.data.table()

#
plot_data <- plot_data[,.(Ratio=value/sum(value),X=X),by=.(group)]
#
ggplot(plot_data, aes(x =group, y= Ratio, fill = X,
                  stratum=X, alluvium=X)) + 
  geom_col(width = 0.6, color='black')+
  geom_flow(width=0.6,alpha=0.3, knot.pos=0)+
  theme_classic() +
  labs(x='Sample',y = 'Ratio')+
  scale_fill_manual(values =   c("#C497B2","#D8383A","#14517C", "#E7EFFA", "#96C37D", "#F3D266", "#2F7FC1", "#F8F3F9","#A9B8C6","#F7E1ED"))+
  scale_x_discrete(expand = c(0.05,0.05))+
  theme(legend.title = element_blank())+
  scale_y_continuous(expand = c(0,0),breaks = c(0.00,0.25,0.5,.75,1),labels = c("0%","25%","50%","75%","100%"))+
  xlab("")+ylab("Cell percentage(%)")+
  theme(axis.text.x =  element_text(size =14,angle = 45,hjust = .9,vjust = .9,color="black"),axis.text.y =  element_text(size =12,color="black"),axis.title = element_text(size = 15))
```
###2.2.2. 
```{r,fig.width=1.5,fig.height=4}
plot_data <- table(sce.big$orig.ident,sce.big$Region)%>%as.data.frame()
colnames(plot_data) <- c("group","X","value")
plot_data$group <- factor(plot_data$group,levels = c("WT","Infect","Infect_IfngrKO"))
plot_data%<>%as.data.table()
#
Pvalue <- lapply(c("Region1","Region2","Region3","Region4","Region5","Region6","Region7","Region8","Region9","Region10"),function(x){
plot_data_sub <- data.frame(
  row.names = c("WT","Infect","Infect_IfngrKO"),
  Region = plot_data[X==x]$value,
  Other = plot_data[X!=x,.(value=sum(value)),by=.(group)]$value
)
a <- chisq.test(plot_data_sub%>%as.matrix())
a <- a$p.value
return(a)
}
)%>%unlist()
p.adjust(Pvalue)

Data <- data.frame(
  Value = -log10(p.adjust(Pvalue)),
  row.names=c("Region1","Region2","Region3","Region4","Region5","Region6","Region7","Region8","Region9","Region10")
)%>%as.matrix()
#
ComplexHeatmap::pheatmap(Data,cluster_rows = F,
                         color = colorRampPalette(rev(brewer.pal(n = 11, name ="RdYlBu")))(100),show_colnames = F,show_rownames = F,
                         annotation_row = data.frame(row.names = c("Region1","Region2","Region3","Region4","Region5","Region6","Region7","Region8","Region9","Region10"),
                                                     Group = c("Region1","Region2","Region3","Region4","Region5","Region6","Region7","Region8","Region9","Region10")
                                                     ),
                         annotation_colors = list(Group =c("Region1"="#C497B2","Region2"="#D8383A","Region3"="#14517C", "Region4"="#E7EFFA","Region5"="#96C37D",
                                                           "Region6"="#F3D266", "Region7"="#2F7FC1","Region8"="#F8F3F9","Region9"="#A9B8C6","Region10"="#F7E1ED"))
                         )
```
#3. DEG analysis of all regions
##3.1. 
```{r,fig.width=5,fig.height=7}
Idents(sce.big) <- "Region"
sce.big %<>% PrepSCTFindMarkers()
DE_ori <- FindAllMarkers(sce.big,only.pos = T,logfc.threshold = 0)
DE_ori%<>%as.data.table()
DE_ori <- subset(DE_ori,p_val_adj<0.05)
```
##3.2.  
```{r}
DE <- lapply(paste0("Region",c(1:10)),function(x){
  Gene <- DE_ori[cluster==x]
  Gene <- Gene$gene
  return(Gene)
}
)
#
names(DE) <- paste0("Region",c(1:10))
Go <- compareCluster(DE,fun="enrichGO", OrgDb='org.Mm.eg.db',keyType = "SYMBOL",ont= "BP")
Go_simple <- clusterProfiler::simplify(Go)
GO_res <- list("Go"=Go,"Go_simple"=Go_simple)
```
##3.3. 
```{r,fig.height=10,fig.width=7}
dotplot(GO_res$Go_simple,showCategory=4,includeAll=F)+
scale_color_gradientn(colours = c("#D6604D","#F4A582","#FDDBC7","#F7F7F7","#D1E5F0","#92C5DE","#4393C3","#2166AC"))+
  theme(axis.text.x = element_text(angle = 45,hjust = .9,vjust = .9))
```
#4. GO enrich of Region2
##4.1. 
```{r,fig.width=5,fig.height=6}
DE <- subset(DE_ori,cluster=="Region2")
Region2_enrich <- enrichGO(DE$gene, OrgDb='org.Mm.eg.db',keyType = "SYMBOL",ont= "BP")
Region2_enrich_simple <- simplify(Region2_enrich)
```
##4.2. 
```{r,fig.width=6,fig.height=4.5}
Region2_enrich_simple@result <- Region2_enrich_simple@result[
  which(Region2_enrich_simple@result$Description%in%c(
    "myeloid leukocyte mediated immunity",
    "positive regulation of adaptive immune response",
    "cell killing",
    "leukocyte cell-cell adhesion","leukocyte migration",
    "leukocyte mediated cytotoxicity",
    "response to interferon-gamma"
  )),]
a <- dotplot(Region2_enrich_simple,size="GeneRatio")+scale_color_gradientn(colours = c("#B2182B","#D6604D","#F4A582","#FDDBC7","#F7F7F7","#D1E5F0","#92C5DE"))
#
ggplot(a$data)+geom_point(aes(x=GeneRatio,y=Description,color=p.adjust,size=GeneRatio))+
  scale_color_gradientn(colours = c("#D6604D","#F4A582","#FDDBC7","#F7F7F7","#D1E5F0","#92C5DE","#4393C3"))+
  theme_bw()+xlab("")+ylab("")+scale_radius(range = c(5,8))
```
